requirements.txt
motor
python-jose
python-dotenv
pydantic
fastapi[all]
uvicorn
bcrypt
passlib
psycopg2
pydantic-settings
pymongo
shortuuid
twilio
cloudinary
sse-starlette
apscheduler

